requirements.txt
amqp==2.4.2
beautifulsoup4==4.7.1
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
cssselect==1.0.3
Django==2.2
django-cors-headers==2.5.2
fakeredis==1.0.3
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.8
jieba3k==0.35.1
kombu==4.5.0
lxml==4.3.3
newspaper3k==0.2.8
nltk==3.4
Pillow==6.0.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
redis==3.2.1
requests==2.21.0
requests-file==1.4.3
singledispatch==3.4.0.3
singleton-decorator==1.0.0
six==1.12.0
sortedcontainers==2.1.0
soupsieve==1.9.1
sqlparse==0.3.0
tinysegmenter==0.3
tldextract==2.2.1
urllib3==1.24.1
vine==1.3.0
